requirements.txt
pywhatkit
pandas
python-dotenv
colorama
loguru
